trusted-types/TrustedTypePolicyFactory-createPolicy.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  async_test(t => {
    let policy = window.trustedTypes.createPolicy('SomeName', s => s )
        .then(t.step_func_done(policy => {
            assert_true(policy instanceof TrustedTypePolicy);
            assert_equals(policy.name, 'SomeName');
    }));
  }, "policy.name = name");

  async_test(t => {
    let policy = window.trustedTypes.createPolicy('SomeName', s => s )
        .then(t.step_func_done(policy => {
            assert_true(policy.createHTML('whatever') instanceof TrustedHTML);
            assert_equals(policy.createHTML('whatever') + "", 'whatever');
    }));
  }, "trustedHTML.html = whatever");
</script>
